test(app): split the app suite along its three seams
880 lines in one file becomes picks, results, and routes, over a shared fixture
module. Same 58 cases and same 99 assertions, moved and not edited.

Each file carries its own vi.mock calls, because those are hoisted per file and
one written in an imported module would not apply.
